+++ src/DataAcquisition/test/uk/ac/ox/zoo/seeg/abraid/mp/dataacquisition/healthmap/HealthMapLookupDataTest.java
@@ -30,19 +30,19 @@ public class HealthMapLookupDataTest {
 
         Country country1 = new Country(1, "Test country 1");
         Country country2 = new Country(2, "Test country 2");
-        HealthMapCountry healthMapCountry1 = new HealthMapCountry(1L, "Test HealthMap country 1", country1);
-        HealthMapCountry healthMapCountry2 = new HealthMapCountry(2L, "Test HealthMap country 2", country2);
+        HealthMapCountry healthMapCountry1 = new HealthMapCountry(1, "Test HealthMap country 1", country1);
+        HealthMapCountry healthMapCountry2 = new HealthMapCountry(2, "Test HealthMap country 2", country2);
 
         List<HealthMapCountry> countries = Arrays.asList(healthMapCountry1, healthMapCountry2);
         when(locationService.getAllHealthMapCountries()).thenReturn(countries);
 
-        Map<Long, HealthMapCountry> expectedCountryMap = new HashMap<>();
-        expectedCountryMap.put(1L, healthMapCountry1);
-        expectedCountryMap.put(2L, healthMapCountry2);
+        Map<Integer, HealthMapCountry> expectedCountryMap = new HashMap<>();
+        expectedCountryMap.put(1, healthMapCountry1);
+        expectedCountryMap.put(2, healthMapCountry2);
 
         // Act
         HealthMapLookupData lookupData = new HealthMapLookupData(alertService, locationService, diseaseService);
-        Map<Long, HealthMapCountry> actualCountryMap = lookupData.getCountryMap();
+        Map<Integer, HealthMapCountry> actualCountryMap = lookupData.getCountryMap();
 
         // Assert
         assertThat(actualCountryMap).isEqualTo(expectedCountryMap);
@@ -57,19 +57,19 @@ public class HealthMapLookupDataTest {
 
         DiseaseGroup disease1 = new DiseaseGroup(1, null, "Test disease 1", DiseaseGroupType.CLUSTER);
         DiseaseGroup disease2 = new DiseaseGroup(2, null, "Test disease 2", DiseaseGroupType.CLUSTER);
-        HealthMapDisease healthMapDisease1 = new HealthMapDisease(1L, "Test HealthMap disease 1", disease1);
-        HealthMapDisease healthMapDisease2 = new HealthMapDisease(2L, "Test HealthMap disease 2", disease2);
+        HealthMapDisease healthMapDisease1 = new HealthMapDisease(1, "Test HealthMap disease 1", disease1);
+        HealthMapDisease healthMapDisease2 = new HealthMapDisease(2, "Test HealthMap disease 2", disease2);
 
         List<HealthMapDisease> diseases = Arrays.asList(healthMapDisease1, healthMapDisease2);
         when(diseaseService.getAllHealthMapDiseases()).thenReturn(diseases);
 
-        Map<Long, HealthMapDisease> expectedDiseaseMap = new HashMap<>();
-        expectedDiseaseMap.put(1L, healthMapDisease1);
-        expectedDiseaseMap.put(2L, healthMapDisease2);
+        Map<Integer, HealthMapDisease> expectedDiseaseMap = new HashMap<>();
+        expectedDiseaseMap.put(1, healthMapDisease1);
+        expectedDiseaseMap.put(2, healthMapDisease2);
 
         // Act
         HealthMapLookupData lookupData = new HealthMapLookupData(alertService, locationService, diseaseService);
-        Map<Long, HealthMapDisease> actualDiseaseMap = lookupData.getDiseaseMap();
+        Map<Integer, HealthMapDisease> actualDiseaseMap = lookupData.getDiseaseMap();
 
         // Assert
         assertThat(actualDiseaseMap).isEqualTo(expectedDiseaseMap);
@@ -83,19 +83,19 @@ public class HealthMapLookupDataTest {
         DiseaseService diseaseService = mock(DiseaseService.class);
 
         Provenance provenance = new Provenance(ProvenanceNames.HEALTHMAP);
-        Feed feed1 = new Feed(10, "Test feed 1", provenance, null, 1, 1L);
-        Feed feed2 = new Feed(20, "Test feed 2", provenance, "de", 1, 2L);
+        Feed feed1 = new Feed(10, "Test feed 1", provenance, null, 1, 1);
+        Feed feed2 = new Feed(20, "Test feed 2", provenance, "de", 1, 2);
 
         List<Feed> feeds = Arrays.asList(feed1, feed2);
         when(alertService.getFeedsByProvenanceName(ProvenanceNames.HEALTHMAP)).thenReturn(feeds);
 
-        Map<Long, Feed> expectedFeedMap = new HashMap<>();
-        expectedFeedMap.put(1L, feed1);
-        expectedFeedMap.put(2L, feed2);
+        Map<Integer, Feed> expectedFeedMap = new HashMap<>();
+        expectedFeedMap.put(1, feed1);
+        expectedFeedMap.put(2, feed2);
 
         // Act
         HealthMapLookupData lookupData = new HealthMapLookupData(alertService, locationService, diseaseService);
-        Map<Long, Feed> actualFeedMap = lookupData.getFeedMap();
+        Map<Integer, Feed> actualFeedMap = lookupData.getFeedMap();
 
         // Assert
         assertThat(actualFeedMap).isEqualTo(expectedFeedMap);
